Listing 26-9. Disabling Debug Mode in the Web.config File
When the debug attribute is set to true, the HTML sent to the browser contains link and script elements for individual files. Whe the attribute is false, the minified versions of the files are selected and concatenated together so that they can be delivered to the client as a blob.